+++ client/src/main/java/com/aapeli/colorgui/Class92.java
@@ -22,7 +22,7 @@ class Class92 implements Runnable {
     }
 
     public void run() {
-        Tools.sleep((long) this.anInt1550);
+        Tools.sleep(this.anInt1550);
         if (this.aBoolean1551) {
             this.aColorSpinner1549.method846();
         }
